Add the Underscore.js library
Underscore is a dependency of Backbone. (Actually, it's the only
dependency of Backbone.)

Once again, I've added both development and minified versions of the
library. The development version is there in case we need to debug or
investigate something; it should be replaced with the minified version
when the app goes into production.