Begin implementing power lines
The idea is that the user will tap a building to begin the power line,
tap zero or more connecting points on the map where pylons will be
placed and finally tap another building to terminate the connection.

So far, only the basic model and 'setter' functions have been
(partially) implemented.