Add a custom BuildingList collection
The collection itself and its buildingAt function, which returns true if
there's a building at the given coordinates (x,y) and false otherwise,
both work.

However, there's a bug present which causes the applications BuildingList*
to always be empty. It's a valid collection, but its length is zero.
Fixing this bug has proved to be tricky.

* app.gameLevel.get("map").get("buildings") in the browser console