Added timer, level and goal alert
There is now added a timer and the
current level on the top of the screen.
The alert is added in the start of every
level. The text in the alert is how much
money the user need to collect in order
to reach the next level. The timer and
the level is not optimal yet, but it is
working. Needs to be added to a own view.