Alerting the goal for the started level
Added a new default value in the model
level called 'goalAlerted'. This is needed
bacause the alert is in the gameloop
and it should only be displayed once.
When alerting the goal, the play mode is
paused and it is started when the player
press ok in the alert box